+++ src/Web/IntersectionObserver/Entry/Safe.js
@@ -1,0 +1,7 @@
+export const time = (entry) => entry.time;
+export const _rootBounds = (entry) => entry.rootBounds;
+export const boundingClientRect = (entry) => entry.boundingClientRect;
+export const intersectionRect = (entry) => entry.intersectionRect;
+export const isIntersecting = (entry) => entry.isIntersecting;
+export const intersectionRatio = (entry) => entry.intersectionRatio;
+export const target = (entry) => entry.target;