+++ src/Web/IntersectionObserver/Entry/Unsafe.js
@@ -1,0 +1,1 @@
+export const _create = (options) => new IntersectionObserverEntry(options);