+++ src/Web/IntersectionObserver/Unsafe.js
@@ -1,0 +1,1 @@
+export const _create = (callback, options) => new IntersectionObserver(callback, options);